Update package version to 1.0.6 and refresh dependencies for improved performance and security
This commit updates the package version to 1.0.6 and refreshes various dependencies in the `package-lock.json` and `package.json` files. Key updates include the upgrade of `@prisma/client`, `next`, and `prisma` to their latest versions, which enhance performance, security, and compatibility. Additionally, a new `logs` model has been added to the Prisma schema, allowing for better tracking of application events. These changes contribute to the overall maintainability and scalability of the application.